refactor(github): relocate GitHubBatchCache + scheduler + controller
Phase 6b of the CQRS refactor (#236, single atomic commit). Moves
the 3 GitHub-cache files from src/modules/execution/ into
src/modules/github/. Their responsibility is caching external GitHub
state — a GitHub concern, not an execution concern. Consolidates
"everything that talks to the gh CLI" in one module so future
consumers (drift reports, planner tools, notifications) can import
GitHubModule instead of reaching into execution internals.

File moves (git mv preserves blame):
  - src/modules/execution/github-batch-cache.service.ts
      → src/modules/github/github-batch-cache.service.ts
  - src/modules/execution/github-cache-scheduler.service.ts
      → src/modules/github/github-cache-scheduler.service.ts
  - src/modules/execution/github-cache.controller.ts
      → src/modules/github/github-cache.controller.ts
  - src/modules/execution/__tests__/github-batch-cache.service.test.ts
      → src/modules/github/__tests__/github-batch-cache.service.test.ts
  - src/modules/execution/__tests__/github-cache-scheduler.test.ts
      → src/modules/github/__tests__/github-cache-scheduler.test.ts

Internal import fixes inside moved files:
  - github-batch-cache.service.ts: ../github/github.service.js
      → ./github.service.js (now same dir)
  - github-cache-scheduler.service.ts: ./events/work-item-merged.event.js
      → ../execution/events/work-item-merged.event.js (event class
      still lives in execution/events/ because that's where it was
      created and it's referenced from other execution services)
  - github-cache-scheduler.test.ts: ../events/... → ../../execution/...

Consumer import updates (3 files in execution/):
  - run-disposition.service.ts, pull-request.service.ts,
    work-item-reconciler.service.ts each swap
    ./github-batch-cache.service.js → ../github/github-batch-cache.service.js

Module wiring:
  - ExecutionModule drops GitHubBatchCache + GitHubCacheScheduler
    from providers + exports, drops GitHubCacheController from
    controllers, drops the three corresponding imports.
  - GitHubModule adds all three to providers (GitHubBatchCache +
    GitHubCacheScheduler), adds GitHubCacheController to controllers,
    adds GitHubBatchCache + GitHubCacheScheduler to exports so the
    execution sub-services can still inject them via the existing
    ExecutionModule → GitHubModule forward-referenced edge.

HTTP route preserved: POST /api/github-cache/refresh still works
because the @Controller("api/github-cache") + @Post("refresh")
decorator values are literal strings that survive the file move.

@Cron(EVERY_MINUTE) on GitHubCacheScheduler.handleTick still fires
because ScheduleModule.forRoot() is wired in app.module.ts; the
scheduler lives in a different module now but the app-level
registration picks up any @Injectable with @Cron decorators.

Same-module cycle note: GitHubBatchCache injects GitHubService (now
in the same module). NestJS handles same-module cycles fine;
GitHubService does not inject GitHubBatchCache back so there's no
real cycle.

forwardRef delta: 0 (stays at 4). The relocation is organizational —
the ExecutionModule → GitHubModule edge was already there (post-6a)
and nothing about this move changes the cycle structure.

All 561 tests pass; tsc --noEmit clean; npm run build:web clean.
